+++ qa/rpc-tests/zcjoinsplit.py
@@ -17,7 +17,7 @@ class JoinSplitTest(BitcoinTestFramework):
     def setup_network(self):
         self.nodes = []
         self.is_network_split = False
-        self.nodes.append(start_node(0, self.options.tmpdir))
+        self.nodes.append(start_node(0, self.options.tmpdir, ["-allowdeprecated=all"]))
 
     def run_test(self):
         zckeypair = self.nodes[0].zcrawkeygen()