+++ qa/rpc-tests/zcjoinsplitdoublespend.py
@@ -7,7 +7,7 @@
 from test_framework.test_framework import BitcoinTestFramework
 from test_framework.authproxy import JSONRPCException
 from test_framework.util import assert_equal, connect_nodes, \
-    gather_inputs, sync_blocks
+    gather_inputs, start_nodes, sync_blocks
 
 import time
 
@@ -15,6 +15,10 @@ class JoinSplitTest(BitcoinTestFramework):
     def setup_network(self):
         # Start with split network:
         return super(JoinSplitTest, self).setup_network(True)
+
+    def setup_nodes(self):
+        return start_nodes(self.num_nodes, self.options.tmpdir,
+                extra_args = [["-allowdeprecated=all"]] * self.num_nodes)
 
     def txid_in_mempool(self, node, txid):
         exception_triggered = False